requirements.txt
networkx
python-igraph
